Add JavaFX.
JavaFX requires the project to have a module-path. =>
Add a module-info.java with all dependencies. This broke the annotation processors. =>
Add processor jars to maven compiler plugin.

Dagger does not define a module, so all its (sub-)dependency jars are automatically named. This breaks the build. =>
Add maven dependency plugin. Override google-java-format (the dependency that imports the build breaking sub-dependency) with the latest version.

Update the Main class in the package plugin and expose the class via module-info.java.